// src/types/generated/MacroState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ObjectiveTimer } from "./ObjectiveTimer";

export type MacroState = { game_time_secs: number, 
/**
 * `GAME_PHASES`.
 */
phase: string, objectives: Array<ObjectiveTimer>, 
/**
 * TR macro reminders for objectives that are up / imminent.
 */
reminders: Array<string>, 
/**
 * TR one-line read of the current macro phase.
 */
phase_note: string, };